--- src/main/frontend/src/store/reducers/requestPopupReducer.js
@@ -1,16 +1,0 @@
-import { handleActions } from 'redux-actions';
-import {
-  openRequestPopup,
-  closeRequestPopup
-} from '../actions/requestPopupAction';
-
-export const initialState = false;
-
-const actionHandlers = {
-  [openRequestPopup]: () => true,
-  [closeRequestPopup]: () => false
-};
-
-const reducer = handleActions(actionHandlers, initialState);
-
-export default reducer;
