+++ src/main/frontend/src/store/selectors.js
@@ -4,5 +4,5 @@ export const getUser = state => state.user;
 export const getCalendar = state => state.calendar;
 export const getPublicHolidays = state => state.publicHolidays;
 export const getEmployees = state => state.employees;
-export const getRequestPopup = state => state.requestPopup;
+export const getRequest = state => state.request;
 export const getLeaveTypes = state => state.leaveTypes;
